setup basic express server to connect to db
seed database with promises library and async queries next!
